backend/requirements.txt
# API and server dependencies
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2

# Google ADK and LLM dependencies
google-adk==0.2.0
google-cloud-aiplatform==1.36.3

# Utility dependencies
python-dotenv==1.0.0
Pillow==10.0.1
requests==2.31.0